Dividing recruiters into the recruiter itself and the service responsible for the API calls.
