+++ dallinger/prolific.py
@@ -1,1 +1,2 @@
-from dallinger.recruiters.prolific import *  # noqa # pylint: disable=unused-import
+from dallinger.recruiters.prolific.recruiter import *  # noqa # pylint: disable=unused-import
+from dallinger.recruiters.prolific.service import *  # noqa # pylint: disable=unused-import
