+++ dallinger/recruiters/service.py
@@ -1,0 +1,4 @@
+class ServiceException(Exception):
+    """Some error from Recruiter API"""
+
+    pass
